Extract view models to their own files
