+++ EssentialFeed/Feed Presentation/FeedErrorViewModel.swift
@@ -1,0 +1,20 @@
+//
+//  FeedErrorViewModel.swift
+//  EssentialFeed
+//
+//  Created by TCode on 21/5/22.
+//
+
+import Foundation
+
+public struct FeedErrorViewModel {
+    public let message: String?
+
+    static var noError: FeedErrorViewModel {
+        return FeedErrorViewModel(message: .none)
+    }
+
+    static func error(message: String) -> FeedErrorViewModel {
+        return FeedErrorViewModel(message: message)
+    }
+}
