+++ EssentialFeed/Feed Presentation/FeedLoadingViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by TCode on 21/5/22.
+//
+
+import Foundation
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
