+++ EssentialFeed/Feed Presentation/FeedViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedViewModel.swift
+//  EssentialFeed
+//
+//  Created by TCode on 21/5/22.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
